pyproject.toml
[build-system]
requires = ["flit_core >=3.9.0,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "dot_tracker"
version = "1.0.0.dev1"
authors = [
    { name = "Roman Averin", email = "mail@romanaverin.com" }
]
description = "A small project to track dotfiles"
readme = "README.md"
license = { "file" = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "PyYAML >= 6.0",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
    "Environment :: Console",
    "Operating System :: POSIX",
    "Topic :: Utilities"]
keywords = ["dotfiles", "manager"]

[project.urls]
"Homepage" = "https://github.com/RomanAverin/dot-tracker"
"Bug Tracker" = "https://github.com/RomanAverin/dot-tracker/issues"


[project.optional-dependencies]
test = [
    "pytest >=7.4.2",
    "tox",
]
# doc = ["sphinx"]

[project.scripts]
dot_tracker = "dot_tracker.main:main"